+++ src/scene_items/color.rs
@@ -1,7 +1,7 @@
 pub struct Color{
-    pub r: f32,
-    pub g: f32,
-    pub b: f32
+    pub r: f64,
+    pub g: f64,
+    pub b: f64
 }
 
 impl Color{
@@ -9,7 +9,7 @@ impl Color{
         Color {r:0.0, g:0.0, b:0.0}
     }
 
-    fn new(r: f32, g: f32, b: f32) -> Color {
+    fn new(r: f64, g: f64, b: f64) -> Color {
         Color{r: r, g: g, b: b}
     }
 
@@ -21,7 +21,7 @@ impl Color{
         Color{r: self.r * other.r, g: self.g * other.g, b: self.b * other.b}
     }
 
-    fn scale(self, factor: f32) -> Color {
+    fn scale(self, factor: f64) -> Color {
         Color{r: self.r * factor, g: self.g * factor, b: self.b * factor}
     }
 }
