+++ src/scene_items/material.rs
@@ -1,6 +1,6 @@
 use scene_items::color::Color;
 pub struct Material{
-    reflection: f32,
+    reflection: f64,
     color: Color
 }
