+++ src/scene_setup/scene.rs
@@ -8,19 +8,19 @@ pub struct Scene{
     pub lights: Vec<Light>
 }
 
-pub fn hit_sphere_ray(sphere: Sphere, ray: Ray, t: i32) -> (bool, f32){
+pub fn hit_sphere_ray(sphere: Sphere, ray: Ray, t: i32) -> (bool, f64){
     let dist = sphere.position.subtract_point(ray.start);
     let b = ray.direction.dot(dist);
     let d = b.pow(2) - dist.dot(dist) + sphere.radius.pow(2);
     if d < 0{
         return (false, -1.0)
     }
-    let t0 = (b as f32) - (d as f32).sqrt();
-    let t1 = (b as f32) + (d as f32).sqrt();
-    if (t1 > 0.1) && (t0 < t as f32) {
+    let t0 = (b as f64) - (d as f64).sqrt();
+    let t1 = (b as f64) + (d as f64).sqrt();
+    if (t1 > 0.1) && (t0 < t as f64) {
         return (true, t0)
     }
-    if (t1 > 0.1) && (t1 < t as f32) {
+    if (t1 > 0.1) && (t1 < t as f64) {
         return (true, t1)
     }
     (false, -1.0)
